requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.6; python_version >= '3.9'
aiohttp==3.11.12; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
aiosqlite==0.20.0; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
async-timeout==4.0.3; python_version < '3.11'
asyncpg==0.30.0; python_full_version >= '3.8.0'
attrs==25.1.0; python_version >= '3.8'
authlib==1.4.1; python_version >= '3.9'
bcrypt==4.2.1
boto3==1.36.21; python_version >= '3.8'
botocore==1.36.21; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
cryptography==44.0.1; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
dnspython==2.7.0; python_version >= '3.9'
email-validator==2.2.0
exceptiongroup==1.2.2; python_version < '3.11'
fastapi[standard]==0.115.8; python_version >= '3.8'
fastapi-cli[standard]==0.0.7
frozenlist==1.5.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httptools==0.6.4
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
iso8601==2.1.0; python_version >= '3.7' and python_version < '4.0'
jinja2==3.1.5
jiter==0.8.2; python_version >= '3.8'
jmespath==1.0.1; python_version >= '3.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
langchain==0.3.18; python_version >= '3.9' and python_version < '4.0'
langchain-aws==0.2.12; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.17; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.35; python_version >= '3.9' and python_version < '4.0'
langchain-openai==0.3.6; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.6; python_version >= '3.9' and python_version < '4.0'
langsmith==0.3.8; python_version >= '3.9' and python_version < '4.0'
loguru==0.7.3; python_version >= '3.5' and python_version < '4.0'
mangum==0.19.0; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.9'
openai==1.63.0; python_version >= '3.8'
orjson==3.10.15; platform_python_implementation != 'PyPy'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
passlib[bcrypt]==1.7.4
propcache==0.2.1; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pydantic-settings==2.7.1; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pypdf==5.3.0; python_version >= '3.8'
pypika-tortoise==0.5.0; python_version >= '3.8' and python_version < '4.0'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==13.9.4; python_full_version >= '3.8.0'
rich-toolkit==0.13.2; python_version >= '3.8'
s3transfer==0.11.2; python_version >= '3.8'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.38; python_version >= '3.7'
starlette==0.45.3; python_version >= '3.9'
tenacity==9.0.0; python_version >= '3.8'
tiktoken==0.9.0; python_version >= '3.9'
tortoise-orm==0.24.0; python_version >= '3.9' and python_version < '4.0'
tqdm==4.67.1; python_version >= '3.7'
typer==0.15.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version < '3.11'
typing-inspect==0.9.0
tzdata==2025.1; python_version >= '2'
urllib3==2.3.0; python_version >= '3.10'
uvicorn[standard]==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
yarl==1.18.3; python_version >= '3.9'
zstandard==0.23.0; python_version >= '3.8'
